web/requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.4.2
Django==1.10.5
django-import-export==1.2.0
django-settings-export==1.2.1
et-xmlfile==1.0.1
flake8==3.2.1
greenlet==0.4.12
gunicorn==19.7.0
jdcal==1.4.1
mccabe==0.5.3
msgpack-python==0.4.8
neovim==0.1.13
odfpy==1.4.0
openpyxl==2.6.2
pbr==1.10.0
psycopg2==2.7
pycodestyle==2.2.0
pyflakes==1.3.0
pyserial==3.2.1
PyYAML==5.1
six==1.10.0
stevedore==1.19.1
tablib==0.13.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
whitenoise==3.3.0
xlrd==1.2.0
xlwt==1.3.0
